Move logout view into CBV
