+++ tests/sentry/web/frontend/accounts/tests.py
@@ -136,24 +136,6 @@ class SettingsTest(TestCase):
         assert user.check_password('foobar')
 
 
-class LogoutTest(TestCase):
-    @fixture
-    def path(self):
-        return reverse('sentry-logout')
-
-    def test_logs_user_out(self):
-        self.login_as(self.user)
-
-        resp = self.client.get(self.path)
-        assert resp.status_code == 302
-        assert self.client.session.keys() == []
-
-    def test_same_behavior_with_anonymous_user(self):
-        resp = self.client.get(self.path)
-        assert resp.status_code == 302
-        assert self.client.session.keys() == []
-
-
 class NotificationSettingsTest(TestCase):
     @fixture
     def path(self):
